# api/openapi.yaml
openapi: 3.1.0
info:
  title: Coop API
  version: 0.1.0
  description: |
    Coop is a self-hosted, parent-curated YouTube for kids.

    This document is the contract between the Go backend and the two native
    clients. The Swift client is generated from it, and the server is tested
    against it so drift fails before release rather than at runtime.

    Two disjoint surfaces share one server:

    - `/api/v1/parent` requires a parent session. Parents scoped to specific
      children receive 404 rather than 403 for children outside their scope, so
      the response does not confirm that another child exists.
    - `/api/v1/child` requires a paired device token. It is read-mostly: a child
      can read their own feed, record local reactions and watch progress, and
      raise a request. Nothing on this surface can change policy.
  license:
    name: MIT
    url: https://github.com/nerdswhofish/coop/blob/main/LICENSE
servers:
  - url: "{publicUrl}/api/v1"
    description: A family's own instance
    variables:
      publicUrl:
        default: https://coop.example
...
security:
  - parentSession: []
paths:
  /setup:
    get:
      operationId: getSetupStatus
      tags: [parent-auth]
      summary: Report whether first-run setup is available
      security: []
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupStatus" }
    post:
      operationId: setUpFamily
      tags: [parent-auth]
      summary: Create the first family and admin
      description: Refused once any family exists on the instance.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "201":
          description: Family created; TOTP enrollment is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthChallenge" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "409":
          description: The instance is already configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /parent/auth/login:
    post:
      operationId: logInParent
      tags: [parent-auth]
      summary: Exchange credentials for a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Password accepted; TOTP verification is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthChallenge" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /parent/auth/invitation:
    post:
      operationId: acceptParentInvitation
      tags: [parent-auth]
      summary: Accept a single-use parent invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, password]
              properties:
                code: { type: string }
                password: { type: string, format: password }
      responses:
        "201":
          description: Parent created; TOTP enrollment is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthChallenge" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /parent/auth/totp/verify:
    post:
      operationId: verifyParentTOTP
      tags: [parent-auth]
      summary: Complete parent authentication with TOTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, code]
              properties:
                challenge: { type: string }
                code: { type: string, pattern: "^[0-9]{6}$" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /parent/me:
    get:
      operationId: getCurrentParent
      tags: [parent-auth]
      summary: The signed-in parent
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Parent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /parent/audit:
    get:
      operationId: listAuditEvents
      tags: [family]
      summary: List retained policy changes visible to this parent
      parameters:
        - name: before
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Newest events first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEvent" }

  # Family -------------------------------------------------------------------
  /parent/family:
    get:
      operationId: getFamily
      tags: [family]
      summary: Read family settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Family" }
    patch:
      operationId: updateFamily
      tags: [family]
      summary: Update family settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                timezone: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Family" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      operationId: deleteFamily
      tags: [family]
      summary: Permanently delete the family and every account and child record
      description: Admin only. This operation cannot be undone.
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
  /parent/family/api-key:
    put:
      operationId: setFamilyAPIKey
      tags: [family]
      summary: Set the family's YouTube Data API key
      description: |
        Admin only. The key is validated against YouTube before it is stored,
        then encrypted at rest. It is never returned by any endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey: { type: string }
      responses:
        "204": { description: Stored }
        "400":
          description: The key was rejected by YouTube
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /parent/parents:
    get:
      operationId: listParents
      tags: [family]
      summary: List parents. Admin only
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Parent" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /parent/parents/invite:
    post:
      operationId: inviteParent
      tags: [family]
      summary: Create a single-use invitation. Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, childIds]
              properties:
                email: { type: string, format: email }
                role:
                  type: string
                  enum: [admin, parent]
                  default: parent
                childIds:
                  type: array
                  description: Children this parent may manage. Must be empty for admins.
                  items: { type: string, format: uuid }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /parent/parents/{parentId}:
    parameters:
      - $ref: "#/components/parameters/ParentId"
    delete:
      operationId: deleteParent
      tags: [family]
      summary: Remove a parent. Admin only
      description: Removing the last admin is refused, since nothing could restore access.
      responses:
        "204": { description: Removed }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Would leave the family with no admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /parent/parents/{parentId}/scope:
    parameters:
      - $ref: "#/components/parameters/ParentId"
    put:
      operationId: setParentScope
      tags: [family]
      summary: Replace which children a parent may manage. Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childIds]
              properties:
                childIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "204": { description: Updated }
        "403": { $ref: "#/components/responses/Forbidden" }

  # Children -----------------------------------------------------------------
  /parent/children:
    get:
      operationId: listChildren
      tags: [children]
      summary: List children in scope
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Child" }
    post:
      operationId: createChild
      tags: [children]
      summary: Create a child profile. Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                avatarId: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /parent/children/{childId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: getChild
      tags: [children]
      summary: Read one child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateChild
      tags: [children]
      summary: Update a child's settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChildSettings" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteChild
      tags: [children]
      summary: Delete a child and everything belonging to them. Admin only
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
  /parent/children/{childId}/recommendations:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: getChildRecommendations
      tags: [children]
      summary: Preview the child's explainable local recommendations
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /parent/children/{childId}/channel-weights:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: getChildChannelWeights
      tags: [children]
      summary: Read non-neutral channel preferences for a child's feed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChannelWeight" }
  /parent/children/{childId}/channel-weights/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/ChannelId"
    put:
      operationId: setChildChannelWeight
      tags: [children]
      summary: Set a soft channel preference without blocking the channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weight]
              properties:
                weight: { type: integer, minimum: -2, maximum: 2 }
      responses:
        "204": { description: Updated }
  /parent/playback:
    get:
      operationId: listActivePlayback
      tags: [child-activity]
      summary: Long-poll the videos children are currently watching
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: Return immediately when omitted; otherwise wait for the active set to change.
      responses:
        "200":
          description: Current playback snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlaybackPage" }
  /parent/children/{childId}/video-blocks:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: listVideoBlocks
      tags: [allowlist]
      summary: Videos explicitly hidden from this child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VideoBlock" }
  /parent/children/{childId}/video-blocks/{videoId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/VideoId"
    put:
      operationId: blockVideoForChild
      tags: [allowlist]
      summary: Hide one video even when its channel is approved
      responses:
        "204": { description: Blocked }
    delete:
      operationId: unblockVideoForChild
      tags: [allowlist]
      summary: Restore normal policy for one video
      responses:
        "204": { description: Unblocked }
        "404": { $ref: "#/components/responses/NotFound" }

  # Pairing ------------------------------------------------------------------
  /parent/children/{childId}/pairing-code:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    post:
      operationId: createPairingCode
      tags: [pairing]
      summary: Mint a single-use pairing code for a child device
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairingCode" }
        "404": { $ref: "#/components/responses/NotFound" }
  /parent/children/{childId}/web-links/{linkId}/approve:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/LinkId"
    post:
      operationId: approveWebLinkAsParent
      tags: [pairing]
      summary: Approve a browser link for a child visible to this parent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebLinkApproval" }
      responses:
        "204": { description: Approved }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /parent/children/{childId}/devices:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: listChildDevices
      tags: [pairing]
      summary: List a child's paired devices
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
  /parent/devices/{deviceId}:
    parameters:
      - name: deviceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: revokeChildDevice
      tags: [pairing]
      summary: Revoke a paired device
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateChildDevice
      tags: [pairing]
      summary: Update controls for a paired device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [allowSelfUnpair]
              properties:
                allowSelfUnpair: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "404": { $ref: "#/components/responses/NotFound" }

  # Allowlists ---------------------------------------------------------------
  /parent/allowlist/global:
    get:
      operationId: getGlobalAllowlist
      tags: [allowlist]
      summary: Channels approved for every child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApprovedChannel" }
    put:
      operationId: allowChannelGlobally
      tags: [allowlist]
      summary: Approve a channel for every child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId: { $ref: "#/components/schemas/ChannelId" }
      responses:
        "204": { description: Approved }
  /parent/allowlist/global/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    delete:
      operationId: disallowChannelGlobally
      tags: [allowlist]
      summary: Remove a channel from the global allowlist
      responses:
        "204": { description: Removed }
  /parent/children/{childId}/allowlist:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      operationId: getChildAllowlist
      tags: [allowlist]
      summary: Channels this child may watch, with the reason for each
      description: |
        Effective state, not a raw table read. Each entry reports whether it
        came from the global list or this child's own list, and whether a
        per-child denial is subtracting it.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EffectiveChannel" }
    put:
      operationId: allowChannelForChild
      tags: [allowlist]
      summary: Approve a channel for this child only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId: { $ref: "#/components/schemas/ChannelId" }
      responses:
        "204": { description: Approved }
  /parent/children/{childId}/allowlist/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/ChannelId"
    delete:
      operationId: disallowChannelForChild
      tags: [allowlist]
      summary: Remove a channel from this child's own allowlist
      responses:
        "204": { description: Removed }
  /parent/children/{childId}/denylist/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/ChannelId"
    put:
      operationId: denyChannelForChild
      tags: [allowlist]
      summary: Subtract a globally approved channel from this child
      responses:
        "204": { description: Denied }
    delete:
      operationId: removeChildChannelDenial
      tags: [allowlist]
      summary: Remove the per-child denial
      responses:
        "204": { description: Removed }
  /parent/blocklist:
    get:
      operationId: getChannelBlocklist
      tags: [allowlist]
      summary: Channels hidden family-wide
      description: Blocked channels are invisible to children and cannot be requested.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlockedChannel" }
    put:
      operationId: blockChannel
      tags: [allowlist]
      summary: Block a channel family-wide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId: { $ref: "#/components/schemas/ChannelId" }
                reason: { type: string }
      responses:
        "204": { description: Blocked }
  /parent/blocklist/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    delete:
      operationId: unblockChannel
      tags: [allowlist]
      summary: Unblock a channel
      responses:
        "204": { description: Unblocked }

  # Keywords -----------------------------------------------------------------
  /parent/keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List negative keywords
      parameters:
        - name: childId
          in: query
          description: Omit for family-wide keywords only.
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Keyword" }
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Create a negative keyword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KeywordInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Keyword" }
...
  /parent/requests:
    get:
      operationId: listParentRequests
      tags: [requests]
      summary: Requests raised by children in scope
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RequestStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestPage" }
...
components:
  securitySchemes:
    parentSession:
      type: http
      scheme: bearer
      description: A parent session token from the login flow.
    childToken:
      type: http
      scheme: bearer
      description: |
        A device token from pairing. Scoped to one child, read-mostly, and
        revocable from the parent app.
    webChildSession:
      type: apiKey
      in: cookie
      name: __Host-coop-child
      description: A Secure, HttpOnly, SameSite=Strict browser device session.
...
  responses:
...
    RateLimited:
      description: Too many attempts
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    SetupStatus:
      type: object
      required: [needsSetup]
      properties:
        needsSetup: { type: boolean }
...